IRAF PCIX Version 2.15.1a
 * 64-bit platform support

+++ unix/hlib/libc/ctype.h
@@ -11,17 +11,17 @@ globalvalue   vms_ctype_defs;
 #endif
 extern	char	u_ctype_[];
 
-#define	isalpha(c)	((u_ctype_+1)[c]&(_U|_L))
-#define	isupper(c)	((u_ctype_+1)[c]&_U)
-#define	islower(c)	((u_ctype_+1)[c]&_L)
-#define	isdigit(c)	((u_ctype_+1)[c]&_N)
-#define	isxdigit(c)	((u_ctype_+1)[c]&(_N|_X))
-#define	isspace(c)	((u_ctype_+1)[c]&_S)
-#define ispunct(c)	((u_ctype_+1)[c]&_P)
-#define isalnum(c)	((u_ctype_+1)[c]&(_U|_L|_N))
-#define isprint(c)	((u_ctype_+1)[c]&(_P|_U|_L|_N))
-#define iscntrl(c)	((u_ctype_+1)[c]&_C)
-#define isascii(c)	((unsigned)(c)<=0177)
+#define	isalpha(c)	((u_ctype_+1)[(unsigned int)(c)]&(_U|_L))
+#define	isupper(c)	((u_ctype_+1)[(unsigned int)(c)]&_U)
+#define	islower(c)	((u_ctype_+1)[(unsigned int)(c)]&_L)
+#define	isdigit(c)	((u_ctype_+1)[(unsigned int)(c)]&_N)
+#define	isxdigit(c)	((u_ctype_+1)[(unsigned int)(c)]&(_N|_X))
+#define	isspace(c)	((u_ctype_+1)[(unsigned int)(c)]&_S)
+#define ispunct(c)	((u_ctype_+1)[(unsigned int)(c)]&_P)
+#define isalnum(c)	((u_ctype_+1)[(unsigned int)(c)]&(_U|_L|_N))
+#define isprint(c)	((u_ctype_+1)[(unsigned int)(c)]&(_P|_U|_L|_N))
+#define iscntrl(c)	((u_ctype_+1)[(unsigned int)(c)]&_C)
+#define isascii(c)	((unsigned)((int)(c))<=0177)
 
 #define toupper(c)	((c)-'a'+'A')
 #define tolower(c)	((c)-'A'+'a')
